refactor: move shared backend to separate file
